refactor(hooks): address PR review feedback
- Rename useGameReplay.ts to gameReplay.ts (not a React hook)
- Remove low-value REPLAY_DELAY_MS constant test
- Fix import ordering for Biome
